Create SmartDevice4.kt
This program implements the use of parameterized constructor(s).
In the SmartDevice class, the name and category properties are immutable. You need to ensure that all the instances of the SmartDevice class initialize the name and category properties. 
With the current implementation, the values for the name and category properties are hard coded. 
This means that all the smart devices are named with the "Android TV" string and categorized with the "Entertainment" string.

In the SmartDevice class, the name and category properties are moved to the constructor without assigning default values by using a parameterized constructor to initialize them. This will maintain immutability and avoid hardcoded values